nit: move interfaces as per review
