+++ src/app/home/tutorial-page/tutorial-page.model.ts
@@ -1,0 +1,10 @@
+export interface ISandbox {
+  terminalUrl: string;
+  websiteUrl: string;
+  arn: string;
+}
+
+export interface IRetryResponse {
+  isUp: boolean;
+  isRetriesExhausted: boolean;
+}
